+++ lib/src/enums.dart
@@ -1,3 +1,5 @@
 /// Enum to define the types of notifications.
 /// Each type has a unique color and icon.
 enum NotificationType { success, error, warning, info }
+
+enum NotificationPosition { topRight, topLeft, bottomRight, bottomLeft }