+++ lib/t_overlay_notification.dart
@@ -1,2 +1,3 @@
+export 'src/enums.dart';
 export 'src/t_notification_overlay.dart';
 export 'src/t_notification_widget.dart';