+++ CHANGELOG.md
@@ -1,5 +1,18 @@
-## 0.0.1
+# Changelog
 
-  Initial release with support for success, error, warning, and info notifications.
-  Stacking notifications are supported.
-  Customizable duration for notification auto-dismiss.
+## 0.0.3
+- Added **LinerProgressIndicator** feature for interactive countdown notifications.
+- Improved **Design** with a more modern, sleek look for notifications.
+- Enhanced positioning options for notifications to support `topLeft`, `topRight`, `bottomLeft`, and `bottomRight`.
+- Introduced **GlobalKey** for better control over notification rendering.
+- Fixed stacking issue and dynamic spacing between notifications.
+- Now you can use **height** variable to adjust height of the notifications.
+
+## 0.0.2
+- Easily use enums to choose **Notification Type**.
+- `NotificationType.info` is now **Default Notification Type**.
+
+## 0.0.1
+- Initial release with support for success, error, warning, and info notifications.
+- **Stacking notifications** are supported.
+- Customizable **duration** for notification auto-dismiss.